:bulb: 카테고리 컬랙션뷰 컴포넌트화
